+++ core/src/com/kikijoli/ville/drawable/entite/decor/Water.java
@@ -7,7 +7,6 @@ package com.kikijoli.ville.drawable.entite.decor;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.kikijoli.ville.manager.ShaderManager;
-import com.kikijoli.ville.maps.Tmap;
 import com.kikijoli.ville.util.Constantes;
 
 /**
@@ -16,7 +15,7 @@ import com.kikijoli.ville.util.Constantes;
  */
 public class Water extends Decor {
 
-    public Water(int srcX, int srcY) {
+    public Water(float srcX, float srcY) {
         super("sprite/water.png", srcX, srcY, Constantes.TILESIZE, Constantes.TILESIZE);
         this.shader = ShaderManager.waveShader;
     }
